+++ CHANGELOG.md
@@ -28,6 +28,8 @@
 * !!! ProcessRepository->countActive()
 * !!! ProcessRepository->getLimitFromItemCountAndOffset()
 * !!! ProcessQueueCommandTest->queueRepository
+* !!! QueueRepository->isPageInQueueTimed()
+* !!! QueueRepository->getAvailableSets()
 
 ## Crawler 11.0.4
 Crawler 11.0.4 was released on February 11th, 2022